+++ classes/Cart.class.php
@@ -180,11 +180,14 @@ class Cart extends Order
         // the item_id.
         // Options are formatted as "id|dscp|price"
         $opts = array();
+        $skus = array($P->name);
         $opt_str = '';          // CSV option numbers
         if (is_array($options) && !empty($options)) {
             foreach($options as $optname=>$option) {
                 $opt_tmp = explode('|', $option);
                 $opts[] = $opt_tmp[0];
+                $Attr = new Attribute($opt_tmp[0]);
+                $skus[] = $Attr->sku;
             }
             $opt_str = implode(',', $opts);
             // Add the option numbers to the item ID to create a new ID
@@ -192,6 +195,8 @@ class Cart extends Order
         } else {
             $options = array();
         }
+
+        $sku = implode('-', $skus);
 
         // Look for identical items, including options (to catch
         // attributes). If found, just update the quantity.