+++ language/english_utf-8.php
@@ -18,6 +18,7 @@ $LANG_SHOP = array (
 'plugin'            => 'Shop',
 'main_title'        => 'Shop',
 'admin_title'       => 'Shop Administration',
+'catalog'           => 'Catalog',
 'blocktitle'        => 'Products',
 'cart_blocktitle'   => 'Shopping Cart',
 'srchtitle'         => 'Products',
@@ -168,7 +169,9 @@ $LANG_SHOP = array (
 'storefront'        => 'Go to Store',
 'options_msg'       => 'Adding options will prevent encrypted buttons from being created.',
 'new_attr'          => 'New Attribute',
+'new_ag'            => 'New Attribute Group',
 'attributes'        => 'Attributes',
+'attr_grp'          => 'Attribute Groups',
 'attr_name'         => 'Attribute Name',
 'attr_value'        => 'Attribute Value',
 'attr_price'        => 'Attribute Price',
@@ -485,6 +488,9 @@ $LANG_SHOP = array (
 'err_missing_email' => 'A valid email address is required.',
 'use_fixed' => 'Include per-item fixed shipping',
 'refunded_x' => 'Refunded %s',
+'first' => 'First',
+'last' => 'Last',
+'pos_after' => 'Position After',
 );
 if (isset($_SHOP_CONF['ena_ratings']) && $_SHOP_CONF['ena_ratings']) {
     $LANG_SHOP['list_sort_options']['top_rated'] = 'Top Rated';
@@ -507,6 +513,7 @@ $LANG_SHOP_HELP = array(
 'sender_name' => 'Optionally enter your name to be shown to the recipient.',
 'gw_bb2_instr' => 'Your IPN URL is %s.<br />Be sure to whitelist it in your Bad Behavior plugin configuration.',
 'gc_message' => 'Enter an optional message to be sent to the recipient (no HTML).',
+'err_sku_nonuniq' => 'There are duplicate SKU values in your product database. Ensure that each product has a unique SKU before using the SKU as the product ID',
 );
 
 $LANG_MYACCOUNT['pe_shop'] = 'Shopping';
@@ -587,6 +594,7 @@ $LANG_confignames['shop'] = array(
     'shop_enabled'  => 'Enable public access?',
     'hp_layout'     => 'Homepage Layout',
     'ipn_url'       => 'IPN URL Override',
+    'sku_url'       => 'Use SKU as Product ID?',
 );
 
 /** Language strings for the subgroup names in the config section */